[Front] refactor(skills): migrate skill configuration GET endpoint to new route
